chore(logger): pass log-mode to each controller and move function to  a dedicated package
- move ConfigLoggers() from common to new logger package
- add new function to convert log-mode string to int for V().
- update README with more detail
