+++ README.md
@@ -228,13 +228,20 @@ This will ensure that the doc for the apis are updated accordingly.
 Logger on all controllers can only be changed from CSV with parameters: --log-mode devel
 valid value: "" (as default) || prod || production || devel || development
 
-This mainly impacts logging for pod startup, generating common resource, monitoring deployment.
+This mainly impacts logging for operator pod startup, generating common resource, monitoring deployment.
+
+| --log-mode value | mapping Log level   | Comments       |
+| ---------------- | ------------------- | -------------- |
+| devel            | debug  / 0          | lowest level   |
+| ""               | info / 1            | default option |
+| default          | info / 1            | default option |
+| prod             | error / 2           | highest level  |
 
 #### Component level
 
 Logger on components can be changed by DSCI devFlags during runtime.
 By default, if not set .spec.devFlags.logmode, it uses INFO level
-Modification applies to all components, not only these "Managed" ones
+Modification applies to all components, not only these "Managed" ones.
 Update DSCI CR with .spec.devFlags.logmode, see example :
 
 ```console
@@ -249,7 +256,15 @@ spec:
 ```
 
 Avaiable value for logmode is "devel", "development", "prod", "production".
-The first two work the same set to DEBUG level; the later two work the same, using ERROR level.
+The first two work the same set to DEBUG level; the later two work the same as using ERROR level.
+
+| .spec.devFlags.logmode | stacktrace level | verbosity | Output   | Comments       |
+| ---------------------- | ---------------- | --------- | -------- | -------------- |
+| devel                  | WARN             | INFO      | Console  | lowest level, using epoch time  |
+| development            | WARN             | INFO      | Console  | same as devel  |
+| ""                     | ERROR            | INFO      | JSON     | default option |
+| prod                   | ERROR            | INFO      | JSON     | highest level, using human readable timestamp  |
+| production             | ERROR            | INFO      | JSON     | same as prod   |
 
 ### Example DSCInitialization
